Refactoring

+++ src/storage/page/mod.rs
@@ -5,8 +5,8 @@ pub use leaf_page::LeafPage;
 pub use internal_page::InternalPage;
 use thiserror::Error;
 
-pub const LEAF_NODE_TAG: u8 = 1;
-pub const INTERNAL_NODE_TAG: u8 = 0;
+pub const LEAF_NODE_TAG: u64 = 1;
+pub const INTERNAL_NODE_TAG: u64 = 0;
 
 #[derive(Debug, Error)]
 pub enum PageCodecError {
